mmdet: r3_cnn original code
