+++ configs/r3_cnn/README.md
@@ -1,0 +1,14 @@
+# Recursively Refined R-CNN: Instance Segmentation with Self-RoI Rebalancing
+
+## Introduction
+
+<!-- [ALGORITHM] -->
+
+```latex
+@article{rossi2021recursively,
+  title={Recursively Refined R-CNN: Instance Segmentation with Self-RoI Rebalancing},
+  author={Rossi, Leonardo and Karimi, Akbar and Prati, Andrea},
+  journal={arXiv preprint arXiv:2104.01329},
+  year={2021}
+}
+```
