+++ mmdet/models/roi_heads/mask_heads/__init__.py
@@ -9,6 +9,7 @@ from .grid_head import GridHead
 from .htc_mask_head import HTCMaskHead
 from .mask_point_head import MaskPointHead
 from .maskiou_head import MaskIoUHead
+from .r3_mask_head import R3MaskHead
 from .scnet_mask_head import SCNetMaskHead
 from .scnet_semantic_head import SCNetSemanticHead
 
@@ -16,5 +17,5 @@ __all__ = [
     'FCNMaskHead', 'HTCMaskHead', 'FusedSemanticHead', 'GridHead',
     'MaskIoUHead', 'CoarseMaskHead', 'MaskPointHead', 'SCNetMaskHead',
     'SCNetSemanticHead', 'GlobalContextHead', 'FeatureRelayHead',
-    'DynamicMaskHead'
+    'DynamicMaskHead', 'R3MaskHead'
 ]
